# fix: improve yarn parser performance by replacing recursion with tail recursion