- [x] Tests written and linted [ℹ︎]()
- [ ] Documentation written [ℹ︎]()
- [x] Commit history is tidy [ℹ︎]()

### What this does

Improve yarn parser performance:

- replacing recursion with tail recursion
- spin event loop every 100 dependencies instead of 200ms

_* I found that major performance issue was nested recursion of `async` function._

### Notes for the reviewer

__Before:__

```
yarn: 233315 deps = 1935.770ms
```

__After:__

```
yarn: 233315 deps = 590.953ms
```

__npm:__

```
npm: 220514 deps = 307.527ms
```

### More information

- https://snyksec.atlassian.net/browse/BST-268
- tested on: https://github.com/snyk-sandbox/big-trees